requirements.txt
django==5.1.1
djangorestframework==3.15.0
psycopg[binary]==3.2.3
celery==5.4.0
redis==5.0.8
dj-database-url==2.2.0
